renamed camelize to transformObject
